Store temperature records in redis
Uses a sorted set to store "event IDs" using the current timestamp as an
integer for the score. Then stores the event data in a hash using the
same "event ID" as the key. This allows for retrieving the event ID keys
in time series order, then fetching the event hash data.